requirements.txt
yandex-speechkit==1.5.0
gigachat==0.1.39
langchain-community==0.0.35
faiss-cpu==1.8.0
sentence-transformers==3.4.1
pandas==2.2.3
sympy==1.13.0
docker==7.0.0
python-dotenv==1.0.1
python-telegram-bot==20.3
python-dotenv==1.0.0
langchain==0.2.2
numpy==1.26.4
faiss-cpu==1.7.4
